Rewrite soup with components

+++ src/main/java/com/github/argon4w/hotpot/client/contents/IHotpotContentRenderer.java
@@ -5,5 +5,5 @@ import com.mojang.blaze3d.vertex.PoseStack;
 import net.minecraft.client.renderer.MultiBufferSource;
 
 public interface IHotpotContentRenderer {
-    void render(IHotpotContent content, PoseStack poseStack, MultiBufferSource bufferSource, int combinedLight, int combinedOverlay, float rotation, float waterLevel, float x, float z);
+    void render(IHotpotContent content, PoseStack poseStack, MultiBufferSource bufferSource, int combinedLight, int combinedOverlay, double rotation, double waterLevel, double x, double z);
 }
