+++ src/main/java/com/github/argon4w/hotpot/client/contents/IHotpotItemContentSpecialRenderer.java
@@ -5,5 +5,5 @@ import com.mojang.blaze3d.vertex.PoseStack;
 import net.minecraft.client.renderer.MultiBufferSource;
 
 public interface IHotpotItemContentSpecialRenderer {
-    void render(AbstractHotpotItemStackContent itemStackContent, PoseStack poseStack, MultiBufferSource bufferSource, int combinedLight, int combinedOverlay, float waterLevel, float rotation, float x, float z);
+    void render(AbstractHotpotItemStackContent itemStackContent, PoseStack poseStack, MultiBufferSource bufferSource, int combinedLight, int combinedOverlay, double waterLevel, double rotation, double x, double z);
 }
