+++ src/main/java/com/github/argon4w/hotpot/client/contents/renderers/HotpotEmptyContentRenderer.java
@@ -7,7 +7,7 @@ import net.minecraft.client.renderer.MultiBufferSource;
 
 public class HotpotEmptyContentRenderer implements IHotpotContentRenderer {
     @Override
-    public void render(IHotpotContent content, PoseStack poseStack, MultiBufferSource bufferSource, int combinedLight, int combinedOverlay, float rotation, float waterLevel, float x, float z) {
+    public void render(IHotpotContent content, PoseStack poseStack, MultiBufferSource bufferSource, int combinedLight, int combinedOverlay, double rotation, double waterLevel, double x, double z) {
 
     }
 }
