+++ src/main/java/com/github/argon4w/hotpot/client/contents/renderers/items/HotpotSkewerItemContentRenderer.java
@@ -11,21 +11,21 @@ import org.joml.Math;
 
 public class HotpotSkewerItemContentRenderer implements IHotpotItemContentSpecialRenderer {
     @Override
-    public void render(AbstractHotpotItemStackContent itemStackContent, PoseStack poseStack, MultiBufferSource bufferSource, int combinedLight, int combinedOverlay, float waterLevel, float rotation, float x, float z) {
+    public void render(AbstractHotpotItemStackContent itemStackContent, PoseStack poseStack, MultiBufferSource bufferSource, int combinedLight, int combinedOverlay, double waterLevel, double rotation, double x, double z) {
         poseStack.pushPose();
 
-        float positionX = 0.5f + x * 0.315f;
-        float positionZ = 0.5f + z * 0.315f;
-        float positionY = 0.56f - getFloatingCurve(rotation / 360.0f, 0.0f) * 0.03f + 0.42f * waterLevel;
+        double positionX = 0.5 + x * 0.315;
+        double positionZ = 0.5 + z * 0.315;
+        double positionY = 0.56 - getFloatingCurve(rotation / 360.0, 0.0f) * 0.03 + 0.42 * waterLevel;
 
-        float degree = - ((float) Math.toDegrees(Math.safeAsin( (waterLevel - 0.35f) / (1.0f - 0.35f))) / 90.0f) * 85.0f;
-        float rotationY = rotation + 20.0f;
-        float rotationX = 270.0f + degree;
+        double degree = - (Math.toDegrees(Math.safeAsin( (waterLevel - 0.35) / (1.0 - 0.35))) / 90.0) * 85.0;
+        double rotationY = rotation + 20.0;
+        double rotationX = 270.0 + degree;
 
         poseStack.translate(positionX, positionY, positionZ);
 
-        poseStack.mulPose(Axis.YP.rotationDegrees(rotationY));
-        poseStack.mulPose(Axis.XN.rotationDegrees(rotationX));
+        poseStack.mulPose(Axis.YP.rotationDegrees((float) rotationY));
+        poseStack.mulPose(Axis.XN.rotationDegrees((float) rotationX));
 
         poseStack.scale(0.32f, 0.32f, 0.32f);
 
@@ -34,7 +34,7 @@ public class HotpotSkewerItemContentRenderer implements IHotpotItemContentSpecia
         poseStack.popPose();
     }
 
-    public float getFloatingCurve(float f, float offset) {
-        return (float) Math.sin((f + offset) / 0.25f * 2f * Math.PI);
+    public double getFloatingCurve(double f, double offset) {
+        return Math.sin((f + offset) / 0.25 * 2 * Math.PI);
     }
 }
